requirements.txt
python-telegram-bot>=22.0,<23.0
httpx>=0.28.0,<1.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.34.0,<1.0.0
python-multipart>=0.0.20,<1.0.0
